first attempt at parsing rasterizer bytecode.
I was able to learn the callback symbol names from the SysV UNIX port of
1-2-3 that included unstripped object files.

This answered a bunch of questions I had, for example, one limitation of
the current ascii-art graphs is that we dont ever see labels. I didn't
really understand why, but now with symbols it's obvious - it was trying
to rasterize them (i.e. convert them into lines), which doesn't show up
on a terminal because they're too fine detail.

With symbols, I was able to figure out how the rasterizer works, 1-2-3
generates little bytecode programs it feeds to the rasterizer engine and
I was able to figure out the opcode names. I think a solution to this
problem is to parse the rasterizer bytecode, figure out when it's trying
to draw text, patch those instructions out and draw themourselves. This
commit is the first stage, I've got parsing kinda working (see raster.c).

It might take some hacking to get this fully working, but it seems
plausible so far!